pagination work

+++ Views/Posts/index.php
@@ -1,3 +1,5 @@
+
+
 <h1>Posts</h1>
 <div class="row col-md-12 centered">
     <table class="table table-striped custab">
@@ -12,23 +14,31 @@
         </thead>
 
         <?php
-
-
-
+        if(!isset($_SESSION)) 
+        { 
+            session_start(); 
+        } 
         foreach ($posts as $post)
         {
             echo '<tr>';
             echo "<td>" . $post['id'] . "</td>";
             echo "<td>" . $post['title'] . "</td>";
             echo "<td>" . $post['description'] . "</td>";
+            if(isset($_SESSION["isLoggedIn"]) && $_SESSION["isLoggedIn"] == "true"){
             echo "<td class='text-center'><a class='btn btn-info btn-xs' href='/blogmvc/posts/edit/" . $post["id"] . "' >
             <span class='glyphicon glyphicon-edit'></span> Edit</a> <a href='/blogmvc/posts/delete/" . $post["id"] . "'
              class='btn btn-danger btn-xs'><span class='glyphicon glyphicon-remove'></span> Del</a></td>";
             echo "</tr>";
+            }
         }
         ?>
     </table>
-
+    
+        <?php
+        foreach($numbers as $num){
+            echo '<a href=" ' . $num . '">'.$num.'</a>';
+            }
+        ?>
 
 
 </div>
